chore: move before/after methods to the test file
