+++ e2e/_suites/fleet/fleet.go
@@ -74,184 +74,6 @@ func (fts *FleetTestSuite) getDeployer() deploy.Deployment {
 	return fts.deployer
 }
 
-// afterScenario destroys the state created by a scenario
-func (fts *FleetTestSuite) afterScenario() {
-	defer func() {
-		fts.DefaultAPIKey = ""
-		// Reset Kibana Profile to default
-		fts.KibanaProfile = ""
-		deployedAgentsCount = 0
-	}()
-
-	span := tx.StartSpan("Clean up", "test.scenario.clean", nil)
-	fts.currentContext = apm.ContextWithSpan(context.Background(), span)
-	defer span.End()
-
-	serviceName := common.ElasticAgentServiceName
-
-	// if DEVELOPER_MODE=true let's not uninstall/unenroll the agent
-	if !common.DeveloperMode && fts.InstallerType != "" {
-		agentService := deploy.NewServiceRequest(serviceName)
-
-		if !fts.StandAlone {
-			// for the centos/debian flavour we need to retrieve the internal log files for the elastic-agent, as they are not
-			// exposed as container logs. For that reason we need to go through the installer abstraction
-			agentInstaller, _ := installer.Attach(fts.currentContext, fts.getDeployer(), agentService, fts.InstallerType)
-
-			if log.IsLevelEnabled(log.DebugLevel) {
-				err := agentInstaller.Logs(fts.currentContext)
-				if err != nil {
-					log.WithField("error", err).Warn("Could not get agent logs in the container")
-				}
-			}
-			// only call it when the elastic-agent is present
-			if !fts.ElasticAgentStopped {
-				err := agentInstaller.Uninstall(fts.currentContext)
-				if err != nil {
-					log.Warnf("Could not uninstall the agent after the scenario: %v", err)
-				}
-			}
-		} else if log.IsLevelEnabled(log.DebugLevel) {
-			// for the Docker image, we simply retrieve container logs
-			_ = fts.getDeployer().Logs(fts.currentContext, agentService)
-		}
-
-		err := fts.unenrollHostname()
-		if err != nil {
-			manifest, _ := fts.getDeployer().Inspect(fts.currentContext, agentService)
-			log.WithFields(log.Fields{
-				"err":      err,
-				"hostname": manifest.Hostname,
-			}).Warn("The agentIDs for the hostname could not be unenrolled")
-		}
-	}
-
-	env := fts.getProfileEnv()
-	_ = fts.getDeployer().Remove(fts.currentContext, deploy.NewServiceRequest(common.FleetProfileName), []deploy.ServiceRequest{deploy.NewServiceRequest(serviceName)}, env)
-
-	// TODO: Determine why this may be empty here before being cleared out
-	if fts.CurrentTokenID != "" {
-		err := fts.kibanaClient.DeleteEnrollmentAPIKey(fts.currentContext, fts.CurrentTokenID)
-		if err != nil {
-			log.WithFields(log.Fields{
-				"err":     err,
-				"tokenID": fts.CurrentTokenID,
-			}).Warn("The enrollment token could not be deleted")
-		}
-	}
-
-	// TODO: Dont think this is needed if we are making all policies unique
-	// fts.kibanaClient.DeleteAllPolicies(fts.currentContext)
-
-	// clean up fields
-	fts.CurrentTokenID = ""
-	fts.CurrentToken = ""
-	fts.InstallerType = ""
-	fts.Image = ""
-	fts.StandAlone = false
-	fts.BeatsProcess = ""
-	fts.ElasticAgentFlags = ""
-}
-
-// beforeScenario creates the state needed by a scenario
-func (fts *FleetTestSuite) beforeScenario() {
-	maxTimeout := time.Duration(utils.TimeoutFactor) * time.Minute
-	exp := utils.GetExponentialBackOff(maxTimeout)
-
-	fts.StandAlone = false
-	fts.ElasticAgentStopped = false
-
-	fts.Version = common.ElasticAgentVersion
-
-	waitForPolicy := func() error {
-		policy, err := fts.kibanaClient.CreatePolicy(fts.currentContext)
-		if err != nil {
-			return errors.Wrap(err, "A new policy could not be obtained, retrying.")
-		}
-
-		log.WithFields(log.Fields{
-			"id":          policy.ID,
-			"name":        policy.Name,
-			"description": policy.Description,
-		}).Info("Policy created")
-
-		fts.Policy = policy
-
-		// Grab the system integration as we'll need to assign it a new name so it wont collide during
-		// multiple policy creations at once
-		integration, err := fts.kibanaClient.GetIntegrationByPackageName(context.Background(), "system")
-		if err != nil {
-			return err
-		}
-
-		packageDataStream := kibana.PackageDataStream{
-			Name:        fmt.Sprintf("%s-%s", integration.Name, uuid.New().String()),
-			Description: integration.Title,
-			Namespace:   "default",
-			PolicyID:    fts.Policy.ID,
-			Enabled:     true,
-			Package:     integration,
-			Inputs:      []kibana.Input{},
-		}
-
-		systemMetricsFile := filepath.Join(testResourcesDir, "/default_system_metrics.json")
-		jsonData, err := readJSONFile(systemMetricsFile)
-		if err != nil {
-			return err
-		}
-
-		for _, item := range jsonData.Children() {
-			var streams []kibana.Stream
-			if err := json.Unmarshal(item.Path("streams").Bytes(), &streams); err != nil {
-				log.WithError(err).Warn("Could not unmarshall streams, will use an empty array instead")
-				streams = []kibana.Stream{}
-			}
-
-			if item.Path("type").Data().(string) == "system/metrics" {
-				packageDataStream.Inputs = append(packageDataStream.Inputs, kibana.Input{
-					Type:    item.Path("type").Data().(string),
-					Enabled: item.Path("enabled").Data().(bool),
-					Streams: streams,
-					Vars: map[string]kibana.Var{
-						"system.hostfs": {
-							Value: "",
-							Type:  "text",
-						},
-					},
-				})
-			} else {
-				packageDataStream.Inputs = append(packageDataStream.Inputs, kibana.Input{
-					Type:    item.Path("type").Data().(string),
-					Enabled: item.Path("enabled").Data().(bool),
-					Streams: streams,
-				})
-			}
-		}
-
-		err = fts.kibanaClient.AddIntegrationToPolicy(context.Background(), packageDataStream)
-		if err != nil {
-			return err
-		}
-
-		return nil
-	}
-
-	err := backoff.Retry(waitForPolicy, exp)
-	if err != nil {
-		log.Fatal(err)
-	}
-
-	// Grab a new enrollment key for new agent
-	enrollmentKey, err := fts.kibanaClient.CreateEnrollmentAPIKey(fts.currentContext, fts.Policy)
-
-	if err != nil {
-		log.Fatal("Unable to create enrollment token for agent")
-	}
-
-	fts.CurrentToken = enrollmentKey.APIKey
-	fts.CurrentTokenID = enrollmentKey.ID
-}
-
 func (fts *FleetTestSuite) anStaleAgentIsDeployedToFleetWithInstaller(staleVersion string, installerType string) error {
 	switch staleVersion {
 	case "latest":
